web-app/projeto_integrador/requirements.txt
asgiref==3.8.1
Django==5.1.8
factory_boy==3.3.3
Faker==37.1.0
gunicorn==23.0.0
packaging==24.1
psycopg==3.2.1
psycopg-binary==3.2.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2025.2
whitenoise==6.7.0
